Separate disassembler functions from display.
Makes it easier to use them in different environments.
